// frontend/src/types/profileSetting.ts
export interface ProfileSetting {
    id: number;
    uuid: string;
    name: string;
    gender: string;
    birth_date: string;
    email: string;
    number_phone: string;
    address: string;
    latitude: string;
    longitude: string;
    city: string;
    created_at: string;
    updated_at: string;
}